test: fix gen-mocks by removing mock_deploy.go
